requirements.txt
Flask==2.2.3
Flask-Cors==3.1.1
Pillow==9.4.0
numpy==1.24.2
scikit-learn==1.2.2
cloudinary==1.30.1
python-dotenv==1.0.0
